# Conflicts:
#	src/main/java/com/avg/lawsuitmanagement/advice/service/AdviceService.java
#	src/main/java/com/avg/lawsuitmanagement/common/exception/type/ErrorCode.java

+++ src/main/java/com/avg/lawsuitmanagement/expense/service/ExpenseService.java
@@ -1,5 +1,9 @@
 package com.avg.lawsuitmanagement.expense.service;
 
+import static com.avg.lawsuitmanagement.common.exception.type.ErrorCode.MEMBER_NOT_ASSIGNED_TO_LAWSUIT;
+
+import com.avg.lawsuitmanagement.common.custom.CustomRuntimeException;
+import com.avg.lawsuitmanagement.common.util.SecurityUtil;
 import com.avg.lawsuitmanagement.expense.controller.form.ExpenseInsertForm;
 import com.avg.lawsuitmanagement.expense.controller.form.ExpenseSearchForm;
 import com.avg.lawsuitmanagement.expense.controller.form.ExpenseUpdateForm;
@@ -8,15 +12,23 @@ import com.avg.lawsuitmanagement.expense.repository.ExpenseMapperRepository;
 import com.avg.lawsuitmanagement.expense.repository.param.ExpenseInsertParam;
 import com.avg.lawsuitmanagement.expense.repository.param.ExpenseSelectParam;
 import com.avg.lawsuitmanagement.expense.repository.param.ExpenseUpdateParam;
+import com.avg.lawsuitmanagement.lawsuit.service.LawsuitService;
+import com.avg.lawsuitmanagement.lawsuit.type.LawsuitStatus;
+import com.avg.lawsuitmanagement.member.repository.MemberMapperRepository;
+import com.avg.lawsuitmanagement.member.service.LoginUserInfoService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
 public class ExpenseService {
 
     private final ExpenseMapperRepository expenseRepository;
+    private final LawsuitService lawsuitService;
+    private final MemberMapperRepository memberMapperRepository;
+    private final LoginUserInfoService loginUserInfoService;
 
     public List<ExpenseDto> searchExpense(ExpenseSearchForm form) {
         return expenseRepository.selectSearchExpense(form.toParam());
@@ -28,17 +40,51 @@ public class ExpenseService {
         return expenseRepository.searchCount(param);
     }
 
+    @Transactional
     public void insertExpense(ExpenseInsertForm form) {
-        ExpenseInsertParam param = form.toParam();
-        expenseRepository.insertExpense(param);
+        // 기존에 등록된 멤버 id 리스트
+        List<Long> originMemberIdList = memberMapperRepository.selectMemberIdListByLawsuitId(form.getLawsuitId());
+
+        if (isUserAuthorizedForLawsuit(loginUserInfoService.getLoginMemberInfo().getId(), originMemberIdList)) {
+            ExpenseInsertParam param = form.toParam();
+
+            expenseRepository.insertExpense(param);
+            lawsuitService.updateStatus(form.getLawsuitId(), LawsuitStatus.PROCEEDING);
+        } else {
+            throw new CustomRuntimeException(MEMBER_NOT_ASSIGNED_TO_LAWSUIT);
+        }
     }
 
     public ExpenseDto updateExpense(Long expenseId, ExpenseUpdateForm form) {
-        expenseRepository.updateExpense(ExpenseUpdateParam.of(expenseId, form));
+        // 기존에 등록된 멤버 id 리스트
+        List<Long> originMemberIdList = memberMapperRepository.selectMemberIdListByLawsuitId(form.getLawsuitId());
+
+        if (isUserAuthorizedForLawsuit(loginUserInfoService.getLoginMemberInfo().getId(), originMemberIdList)) {
+            expenseRepository.updateExpense(ExpenseUpdateParam.of(expenseId, form));
+        } else {
+            throw new CustomRuntimeException(MEMBER_NOT_ASSIGNED_TO_LAWSUIT);
+        }
+
         return expenseRepository.selectExpenseById(expenseId);
     }
 
-    public void deleteExpense(Long expenseId) {
-        expenseRepository.deleteExpense(expenseId);
+    public void deleteExpense(Long expenseId, Long lawsuitId) {
+        // 기존에 등록된 멤버 id 리스트
+        List<Long> originMemberIdList = memberMapperRepository.selectMemberIdListByLawsuitId(lawsuitId);
+
+        if (isUserAuthorizedForLawsuit(loginUserInfoService.getLoginMemberInfo().getId(), originMemberIdList)) {
+            expenseRepository.deleteExpense(expenseId);
+        } else {
+            throw new CustomRuntimeException(MEMBER_NOT_ASSIGNED_TO_LAWSUIT);
+        }
+    }
+
+    private boolean isUserAuthorizedForLawsuit(long userId, List<Long> memberIds) {
+        if (SecurityUtil.getCurrentLoginRoleList().contains("ROLE_ADMIN")) {
+            return true;
+        }
+
+        // 로그인한 사용자가 해당 사건의 담당자가 아니라면
+        return memberIds.contains(userId);
     }
 }
